Working FB + discrete gridworld + continuous + hydra expertiments

+++ envs/custom_mazes/generators/four_room.py
@@ -1,8 +1,7 @@
 import numpy as np
-from skimage.draw import rectangle
 
 def generate_four_room_env(width, height, number_of_space_between_wall=3):
-    maze = np.zeros((height, width), dtype=np.uint8)
+    maze = np.zeros((height, width), dtype=np.int32)
     maze[0, :] = 1 # up
     maze[-1, :] = 1 # down
     maze[:, 0] = 1 # left
@@ -17,4 +16,5 @@ def generate_four_room_env(width, height, number_of_space_between_wall=3):
     return maze
     # rr, cc = rectangle([1, 1], extent=[thick, t_shape[0] + thick])
     # x[rr, cc] = 0
-     
+    
+
